Splitting reducers
